pyproject.toml
[project]
name = "smoke-signal"
version = "0.2.0"
description = "Smoke Signal — local-first auto-transcription with speaker diarization"
requires-python = ">=3.11,<3.13"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "torch>=2.10.0",
    "whisperx>=3.3",
    "pyannote.audio>=4.0",
    "soundfile>=0.12",
    "numpy>=1.26",
]

[project.scripts]
smoke-signal = "smoke_signal.cli:main"

[project.gui-scripts]
smoke-signal-tray = "smoke_signal.tray_entry:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
smoke_signal = ["config.yaml.example"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
watch = [
    "watchdog>=4.0",
    "winotify>=1.1; sys_platform == 'win32'",
    "pystray>=0.19",
    "Pillow>=10.0",
]
